% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstndev-package.r
\docType{package}
\name{rstndev}
\alias{rstndev}
\alias{rstndev-package}
\title{rstndev: Calculates Standard Deviation}
\description{
Calculating standard error by hand can be difficult.
If for some reason you would like a whole separate package for this function,
this package will take in a number and return the standard deviation of
that number.
}

