% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stndevscript.R
\name{standard_error}
\alias{standard_error}
\title{Calculate Standard Error}
\usage{
standard_error(x)
}
\arguments{
\item{x}{as a list}
}
\value{
numerical
}
\description{
Returns the standard error of some given values
}
\examples{
#' standard_error(c(1,2,3,4))
}

