% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stndevscript.R
\name{standard_deviation}
\alias{standard_deviation}
\title{Calculate Standard Deviation}
\usage{
standard_deviation(x)
}
\arguments{
\item{x}{as a list}
}
\value{
numerical
}
\description{
Returns the standard deviation of some given values
}
\examples{
#' standard_deviation(c(1,2,3,4))
}

